// src/models/IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceConversion/types.ts
import { IoK8sApiextensionsApiserverPkgApisApiextensionsV1WebhookConversionModelOptions as IoK8sApiextensionsApiserverPkgApisApiextensionsV1WebhookConversion } from '../IoK8sApiextensionsApiserverPkgApisApiextensionsV1WebhookConversion';

export interface IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceConversionModelOptions {
  strategy: string;
  webhook?: IoK8sApiextensionsApiserverPkgApisApiextensionsV1WebhookConversion;
}